Address PR review feedback
- Split commands.ts into commands/{leaf,composed,pack,index}.ts
- Rename bin.ts to main.ts (matches default entryPoints pattern)
- Revert eslint.config.ts override (no longer needed)
- Slim root scripts to CI/lifecycle needs only (gtb, prepare,
  build:ci, prepack, test:e2e); use pnpm run gtb directly
- Update AGENTS.md and README.md for direct gtb usage
