+++ eslint.config.ts
@@ -1,6 +1,5 @@
 import { configure } from '@gtbuchanan/eslint-config';
 
 export default configure({
-  entryPoints: ['**/bin/**/*.ts', '**/bin.ts', '**/main.ts'],
   tsconfigRootDir: import.meta.dirname,
 });
